% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appendix_functions.R
\name{b_functionA3}
\alias{b_functionA3}
\title{B function from Proposition A3}
\usage{
b_functionA3(obs_draws, g, psi)
}
\arguments{
\item{obs_draws}{Row of the data.frame of observable draws}

\item{g}{Value from g function}

\item{psi}{Psi value}
}
\value{
A min and a max of the B function
}
\description{
B function from Proposition A3
}
