% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colonial}
\alias{colonial}
\title{Acemoglu, Johnson, and Robinson (2001) Dataset}
\format{
A data frame with 64 rows and 9 variables:
\describe{
   \item{shortnam}{three letter country abbreviation, e.g. AUS for Australia}
   \item{africa}{dummy variable =1 if country is in Africa}
   \item{lat_abst}{absolute distance to equator (scaled between 0 and 1)}
   \item{rich4}{dummy variable, =1 for "Neo-Europes" (AUS, CAN, NZL, USA)}
   \item{avexpr}{Average protection against expropriation risk.
   Measures risk of government appropriation of foreign private investment
   on a scale from 0 (least risk) to 10 (most risk). Averaged over all years
   from 1985-1995.}
   \item{logpgp95}{Natural logarithm of per capita GDP in 1995 at purchasing
   power parity}
   \item{logem4}{Natural logarithm of European settler mortality}
   \item{asia}{dummy variable, =1 if country is in Asia}
   \item{loghjypl}{Natural logarithm of output per worker in 1988}
 }
}
\source{
\url{http://economics.mit.edu/faculty/acemoglu/data/ajr2001}
}
\usage{
colonial
}
\description{
Cross-country dataset used to construct Table 4 of Acemoglu, Johnson & Robinson (2001).
}
\references{
\url{https://www.aeaweb.org/articles.php?doi=10.1257/aer.91.5.1369}
}
\keyword{datasets}
