% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{covCLT}
\alias{covCLT}
\title{Posterior draws for IV and OLS based on a large-sample approximation}
\usage{
covCLT(inData, n_draws)
}
\arguments{
\item{inData}{A dataframe with columns named x, y, and z each of which is a
numeric vector.}

\item{n_draws}{An integer, the number of posterior draws.}
}
\value{
An array of numeric matrices, each of which is a sample from the
marginal posterior of the covariance matrix.
}
\description{
Posterior draws for IV and OLS based on a large-sample approximation
}
\details{
In OLS regression we condition on x and in IV regression we
condition on x and z. This function generates posterior draws for the
covariance matrix of (x, y, z) conditional on the observed values of x and
z based on the usual large-sample frequentist asymptotics. In particular
from the CLT, the OLS and IV estimators are jointly asymptotically normal.
Treating this asymptotic result as exact and multiplying through by the
sample estimates of Var(x) and Cov(x,z) gives a joint normal distribution for
Cov(x,y) and Cov(z,y). To get draws for Cov(x,y,z) we simply fill in the
sample estimates of Var(x), Var(y), Var(z), and Cov(x,z). When generating
draws for Cov(x,y) and Cov(z,y) we allow for heteroskedasticity. Note that
when generated in this way, posterior draws are not guaranteed to be positive
definite. If any of the draws is not positive definite the function throws
an error.
}
\examples{
set.seed(1234)
x <- rnorm(100)
y <- x / 3 + rnorm(100)
z <- y / 5 + rnorm(100)
covCLT(data.frame(x, y, z), 10)
}

