% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{covJeffreys}
\alias{covJeffreys}
\title{Posterior draws for a cov matrix based on Jeffrey's prior}
\usage{
covJeffreys(inData, n_draws)
}
\arguments{
\item{inData}{A matrix or dataframe, each column of which is a variable and
each row an observation.}

\item{n_draws}{An integer, the number of posterior draws.}
}
\value{
An array of numeric matrices, each of which is a sample from the
marginal posterior of the covariance matrix.
}
\description{
Posterior draws for a cov matrix based on Jeffrey's prior
}
\details{
Based on a joint normal likelihood, not conditional as in
regression.
}
\examples{
set.seed(1234)
x <- rnorm(100)
y <- x / 3 + rnorm(100)
z <- y / 5 + rnorm(100)
covJeffreys(cbind(x, y, z), 10)
}

