% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{devech}
\alias{devech}
\title{Convert Half-vectorization to symmetric matrix}
\usage{
devech(v, dim)
}
\arguments{
\item{v}{A numeric vector with n * (n + 1) / 2 elements.}

\item{dim}{An integer indicating the dimension of the resulting square,
symmetric matrix.}
}
\value{
The n by n symmetric matrix whose half-vectorization is v.
}
\description{
Convert Half-vectorization to symmetric matrix
}
\details{
Throws an error if dim does not correspond to the length of v.
}
\examples{
v <- c(11:14, 22:24, 33:34, 44)
devech(v, 4)
}

