% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{draw_observables}
\alias{draw_observables}
\title{Simulates different data draws}
\usage{
draw_observables(y_name, T_name, z_name, data, controls = NULL, n_draws = 5000)
}
\arguments{
\item{y_name}{Character vector of the name of the dependent variable}

\item{T_name}{Character vector of the names of the preferred regressors}

\item{z_name}{Character vector of the names of the instrumental variables}

\item{data}{Data to be analyzed}

\item{controls}{Character vector containing the names of the exogenous regressors}

\item{n_draws}{Integer number of simulations to draw}
}
\value{
Data frame containing covariances, correlations, and R-squares for
  each data simulation
}
\description{
This function takes the data and simulates potential draws of data from
  the properties of the observed data.
}
