% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{draw_sigma_CLT}
\alias{draw_sigma_CLT}
\title{Draws covariance matrix using the CLT}
\usage{
draw_sigma_CLT(y, Tobs, z, n_draws)
}
\arguments{
\item{y}{Vector of dependent variable data}

\item{Tobs}{Matrix containing data for the endogenous regressor}

\item{z}{Matrix containing data for the instrumental variable}

\item{n_draws}{Integer number of draws to generate}
}
\value{
List of covariance matrix draws (Sigma) and an indicator of whether
  each one is positive definite
}
\description{
This function takes the data and assumes that the covariance between the
  dependent variable and the preferred regressor (Tobs and y) and the
  covariance between the instrument and the dependent variable is normal.
  However, only drawing these parameters does not guarantee that the
  full covariance matrix will be positive definite. However, this does
  avoid the necessity of imposing a prior and redrawing the whole covariance
  matrix using Bayesian methods.
}
