% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingUncertainty.R
\name{draw_sigma_jeffreys}
\alias{draw_sigma_jeffreys}
\title{Draws covariance matrix using the Jeffrey's Prior}
\usage{
draw_sigma_jeffreys(y, Tobs, z, k, n_draws)
}
\arguments{
\item{y}{Vector of dependent variable}

\item{Tobs}{Matrix containing data for the preferred regressor}

\item{z}{Matrix containing data for the instrumental variable}

\item{k}{Number of covariates, including the intercept}

\item{n_draws}{Integer number of draws to perform}
}
\value{
Array of covariance matrix draws
}
\description{
Draws covariance matrix using the Jeffrey's Prior
}
