% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{format_est}
\alias{format_est}
\title{Creates LaTeX code for parameter estimates}
\usage{
format_est(est)
}
\arguments{
\item{est}{Number}
}
\value{
LaTeX string for the number
}
\description{
Creates LaTeX code for parameter estimates
}
