% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{format_HPDI}
\alias{format_HPDI}
\title{Creates LaTeX code for the HPDI}
\usage{
format_HPDI(bounds)
}
\arguments{
\item{bounds}{2-element vector of the upper and lower HPDI bounds}
}
\value{
LaTeX string of the HPDI
}
\description{
Creates LaTeX code for the HPDI
}
