% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{format_se}
\alias{format_se}
\title{Creates LaTeX code for the standard error}
\usage{
format_se(se)
}
\arguments{
\item{se}{Standard error}
}
\value{
LaTeX string for the standard error
}
\description{
Creates LaTeX code for the standard error
}
