% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appendix_functions.R
\name{g_functionA2}
\alias{g_functionA2}
\title{G function from Proposition A.2}
\usage{
g_functionA2(kappa, r_TstarU, obs_draws)
}
\arguments{
\item{kappa}{Kappa value}

\item{r_TstarU}{r_TstarU value}

\item{obs_draws}{a row of the data.frame of observable draws}
}
\value{
G value
}
\description{
G function from Proposition A.2
}
