% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_bounds.R
\name{get_alpha_bounds}
\alias{get_alpha_bounds}
\title{Computes a0 and a1 bounds}
\usage{
get_alpha_bounds(draws, p)
}
\arguments{
\item{draws}{data.frame of observables of simulated data}

\item{p}{Treatment probability from binary data}
}
\value{
List of alpha bounds
}
\description{
Computes a0 and a1 bounds
}
