% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appendix_functions.R
\name{get_beta_bounds_binary_post}
\alias{get_beta_bounds_binary_post}
\title{Generates beta bounds off of beta draws}
\usage{
get_beta_bounds_binary_post(draws, n_observables)
}
\arguments{
\item{draws}{Posterior draws}

\item{n_observables}{Number of observable draws}
}
\value{
Upper and lower bounds of beta based on posterior draws
}
\description{
Generates beta bounds off of beta draws
}
