% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_bounds_unrest}
\alias{get_bounds_unrest}
\title{Wrapper function combines all unrestricted bounds together. Vectorized}
\usage{
get_bounds_unrest(obs)
}
\arguments{
\item{obs}{Observables generated by get_observables}
}
\value{
List of unrestricted bounds for r_TstarU, r_uz, and kappa
}
\description{
Wrapper function combines all unrestricted bounds together. Vectorized
}
