% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.R
\name{get_estimates}
\alias{get_estimates}
\title{Computes OLS and IV estimates}
\usage{
get_estimates(y_name, T_name, z_name, data, controls = NULL, robust = FALSE)
}
\arguments{
\item{y_name}{Character vector of the name of the dependent variable}

\item{T_name}{Character vector of the names of the preferred regressors}

\item{z_name}{Character vector of the names of the instrumental variables}

\item{data}{Data to be analyzed}

\item{controls}{Character vector containing the names of the exogenous regressors}

\item{robust}{Boolean of whether to compute heteroskedasticity-robust standard errors}
}
\value{
List of beta estimates and associated standard errors for OLS and
  IV estimation
}
\description{
Computes OLS and IV estimates
}
