% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_k_bounds_unrest}
\alias{get_k_bounds_unrest}
\title{Given observables from the data, generates unrestricted bounds for kappa.
  Vectorized}
\usage{
get_k_bounds_unrest(obs, tilde)
}
\arguments{
\item{obs}{Observables generated by get_observables}

\item{tilde}{Boolean of whether or not kappa_tilde or kappa is desired}
}
\value{
List of upper bounds and lower bounds for kappa
}
\description{
Given observables from the data, generates unrestricted bounds for kappa.
  Vectorized
}
