% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_bounds.R
\name{get_L}
\alias{get_L}
\title{Computes L, lower bound for kappa_tilde in paper}
\usage{
get_L(draws)
}
\arguments{
\item{draws}{data.frame of observables of simulated data}
}
\value{
Vector of L values
}
\description{
Computes L, lower bound for kappa_tilde in paper
}
