% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_M}
\alias{get_M}
\title{Solves for the magnification factor}
\usage{
get_M(r_TstarU, k, obs)
}
\arguments{
\item{r_TstarU}{Vector of r_TstarU values}

\item{k}{Vector of kappa values}

\item{obs}{Observables generated by get_observables}
}
\value{
Vector of magnification factors
}
\description{
This function solves for the magnification factor given r_TstarU and kappa.
  It handles 3 potential cases when the magnification factor must be evaluated:
  1. Across multiple simulations, but given the same r_TstarU and k
  2. For multiple simulations, each with a value of r_TstarU and k
  3. For one simulation across a grid of r_TstarU and k
}
