% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valid_IV_Beliefs.R
\name{get_new_draws}
\alias{get_new_draws}
\title{Computes beliefs that support valid instrument}
\usage{
get_new_draws(obs_draws, post_draws)
}
\arguments{
\item{obs_draws}{data.frame of draws of reduced form parameters}

\item{post_draws}{data.frame of posterior draws}
}
\value{
data.frame of new draws
}
\description{
Computes beliefs that support valid instrument
}
