% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_observables}
\alias{get_observables}
\title{Given data and function specification, returns the relevant correlations and
  covariances with any exogenous controls projected out.}
\usage{
get_observables(y_name, T_name, z_name, data, controls = NULL)
}
\arguments{
\item{y_name}{Name of the dependent variable}

\item{T_name}{Name(s) of the preferred regressor(s)}

\item{z_name}{Name(s) of the instrumental variable(s)}

\item{data}{Data to be analyzed}

\item{controls}{Exogenous regressors to be included}
}
\value{
List of correlations, covariances, and R^2 of first and second stage
  regressions after projecting out any exogenous control regressors
}
\description{
Given data and function specification, returns the relevant correlations and
  covariances with any exogenous controls projected out.
}
