% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_p_valid}
\alias{get_p_valid}
\title{Compute the share of draws that could contain a valid instrument.}
\usage{
get_p_valid(draws)
}
\arguments{
\item{draws}{List of simulated draws}
}
\value{
Numeric of the share of valid draws
}
\description{
Compute the share of draws that could contain a valid instrument.
}

