% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_r_TstarU_bounds_unrest}
\alias{get_r_TstarU_bounds_unrest}
\title{Given observables from the data, generates the unrestricted bounds for rho_TstarU.
  Data does not impose any restrictions on r_TstarU
  Vectorized}
\usage{
get_r_TstarU_bounds_unrest(obs)
}
\arguments{
\item{obs}{Observables generated by get_observables}
}
\value{
List of upper and lower bounds for r_TstarU
}
\description{
Given observables from the data, generates the unrestricted bounds for rho_TstarU.
  Data does not impose any restrictions on r_TstarU
  Vectorized
}
