% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rho_bounds.R
\name{get_r_uz_bounds}
\alias{get_r_uz_bounds}
\title{Evaluates r_uz bounds given user restrictions on r_TstarU and kappa}
\usage{
get_r_uz_bounds(r_TstarU_lower, r_TstarU_upper, k_lower, k_upper, obs)
}
\arguments{
\item{r_TstarU_lower}{Vector of lower bounds of endogeneity}

\item{r_TstarU_upper}{Vector of upper bounds of endogeneity}

\item{k_lower}{Vector of lower bounds on measurement error}

\item{k_upper}{Vector of upper bounds on measurement error}

\item{obs}{Observables generated by get_observables}
}
\value{
2-column data frame of lower and upper bounds of r_uz
}
\description{
This function takes observables from the data and user beliefs over the extent
  of measurement error (kappa) and the direction of endogeneity (r_TstarU) to
  generate the implied bounds on instrument validity (r_uz)
}
