% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedSet.R
\name{get_r_uz_bounds_unrest}
\alias{get_r_uz_bounds_unrest}
\title{Given observables from the data, generates the unrestricted bounds for rho_uz.
  Vectorized}
\usage{
get_r_uz_bounds_unrest(obs)
}
\arguments{
\item{obs}{Observables generated by get_observables}
}
\value{
List of upper and lower bounds for rho_uz
}
\description{
Given observables from the data, generates the unrestricted bounds for rho_uz.
  Vectorized
}
