% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverage.R
\name{getCoverage}
\alias{getCoverage}
\title{Computes coverage of list of intervals}
\usage{
getCoverage(data, guess)
}
\arguments{
\item{data}{2-column data frame of confidence intervals}

\item{guess}{2-element vector of confidence interval}
}
\value{
Coverage percentage
}
\description{
Computes coverage of list of intervals
}
