% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoverage.R
\name{getInterval}
\alias{getInterval}
\title{Generates smallest covering interval}
\usage{
getInterval(data, center, conf = 0.9, tol = 1e-06)
}
\arguments{
\item{data}{2-column data frame of confidence intervals}

\item{center}{2-element vector to center coverage interval}

\item{conf}{Confidence level}

\item{tol}{Tolerance level for convergence}
}
\value{
2-element vector of confidence interval
}
\description{
Generates smallest covering interval
}
