% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{make_full_row}
\alias{make_full_row}
\title{Takes the OLS and IV estimates and converts it to a row of the LaTeX table}
\usage{
make_full_row(stats, example_name)
}
\arguments{
\item{stats}{List with OLS and IV estimates and the bounds on kappa and r_uz}

\item{example_name}{Character string detailing the example}
}
\value{
LaTeX code passed to makeTable()
}
\description{
Takes the OLS and IV estimates and converts it to a row of the LaTeX table
}
