% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{make_tex_row}
\alias{make_tex_row}
\title{Makes LaTeX code to make a row of a table and shift by some amount of columns
  if necessary}
\usage{
make_tex_row(char_vec, shift = 0)
}
\arguments{
\item{char_vec}{Vector of characters to be collapsed into a LaTeX table}

\item{shift}{Number of columns to shift over}
}
\value{
LaTeX string of the whole row of the table
}
\description{
Makes LaTeX code to make a row of a table and shift by some amount of columns
  if necessary
}
