% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{makeExample}
\alias{makeExample}
\title{Generates table of parameter estimates given user restrictions and data}
\usage{
makeExample(
  y_name,
  T_name,
  z_name,
  data,
  controls = NULL,
  robust = FALSE,
  r_TstarU_restriction = NULL,
  k_restriction = NULL,
  n_draws = 5000,
  n_RF_draws = 1000,
  n_IS_draws = 1000,
  resample = FALSE,
  example_name
)
}
\arguments{
\item{y_name}{Character string with the column name of the dependent variable}

\item{T_name}{Character string with the column name of the endogenous regressor(s)}

\item{z_name}{Character string with the column name of the instrument(s)}

\item{data}{Data frame}

\item{controls}{Vector of character strings specifying the exogenous variables}

\item{robust}{Indicator for heteroskedasticity-robust standard errors}

\item{r_TstarU_restriction}{Matrix of desired mins and maxes for r_TstarU (must be same dimensions as k_restriction). If NULL, defaults to [-0.999, 0.999]}

\item{k_restriction}{Matrix of desired minx and maxes for kappa (must be same dimensions as r_TstarU_restriction). If NULL, defaults to [0.001, 0.999]}

\item{n_draws}{Number of draws when generating frequentist-friendly draws of the covariance matrix}

\item{n_RF_draws}{Number of reduced-form draws}

\item{n_IS_draws}{Number of draws on the identified set}

\item{resample}{Indicator of whether or not to resample using magnification factor}

\item{example_name}{Character string describing the example}
}
\value{
LaTeX code to be passed to makeTable() function
}
\description{
Generates table of parameter estimates given user restrictions and data
}
