% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{makeTable}
\alias{makeTable}
\title{Generates table of parameter estimates given user restrictions and data}
\usage{
makeTable(file, binary, ...)
}
\arguments{
\item{file}{Character string with the path name of the .tex file to be saved}

\item{binary}{Indicator for if data is binary}

\item{...}{Arguments of TeX code for individual examples to be combined into a single table}
}
\value{
LaTeX code that generates output table with regression results
}
\description{
Generates table of parameter estimates given user restrictions and data
}
