% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plot.R
\name{map2color}
\alias{map2color}
\title{Generates a custom color palette given a vector of numbers}
\usage{
map2color(x, pal, limits = NULL)
}
\arguments{
\item{x}{Vector of numbers}

\item{pal}{Palette function generate from colorRampPalette}

\item{limits}{Limits on the numeric sequence}
}
\value{
Hex values for colors
}
\description{
Generates a custom color palette given a vector of numbers
}
