% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_functions.R
\name{myformat}
\alias{myformat}
\title{Rounds x to two decimal places}
\usage{
myformat(x)
}
\arguments{
\item{x}{Number to be rounded}
}
\value{
Number rounded to 2 decimal places
}
\description{
Rounds x to two decimal places
}
