% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plot.R
\name{plot_3d_beta}
\alias{plot_3d_beta}
\title{Plot ivdoctr Restrictions}
\usage{
plot_3d_beta(
  y_name,
  T_name,
  z_name,
  data,
  controls = NULL,
  r_TstarU_restriction = NULL,
  k_restriction = NULL,
  n_grid = 30,
  n_colors = 500,
  fence = NULL,
  gray_k = NULL,
  gray_rTstarU = NULL,
  theta = 0,
  phi = 15
)
}
\arguments{
\item{y_name}{Character string with the column name of the dependent variable}

\item{T_name}{Character string with the column name of the endogenous regressor(s)}

\item{z_name}{Character string with the column name of the instrument(s)}

\item{data}{Data frame}

\item{controls}{Vector of character strings specifying the exogenous variables}

\item{r_TstarU_restriction}{Matrix of desired mins and maxes for r_TstarU (must be same dimensions as k_restriction)}

\item{k_restriction}{Matrix of desired minx and maxes for kappa (must be same dimensions as r_TstarU_restriction)}

\item{n_grid}{Number of points to put in grid}

\item{n_colors}{Number of colors to use}

\item{fence}{Vector of left, bottom, right, and top corners of rectangle}

\item{gray_k}{Matrix of kappa restrictions to recolor graph as gray}

\item{gray_rTstarU}{Matrix of rTstarU restrictions to recolor graph as gray}

\item{theta}{Graphing parameters for orienting plot}

\item{phi}{Graphing parameters for orienting plot}
}
\value{
Interactive 3d plot which can be oriented and saved using rgl.snapshot()
}
\description{
Plot ivdoctr Restrictions
}
