% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rect_points}
\alias{rect_points}
\title{Construct vectors of points that outline a rectangle.}
\usage{
rect_points(xleft, ybottom, xright, ytop, step_x, step_y)
}
\arguments{
\item{step_y}{}
}
\description{
Construct vectors of points that outline a rectangle.
}

