% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvwish}
\alias{rinvwish}
\title{Simulate draws from the inverse Wishart distribution}
\usage{
rinvwish(n, v, S)
}
\arguments{
\item{n}{An integer, the number of draws.}

\item{v}{An integer, the degrees of freedom of the distribution.}

\item{S}{A numeric matrix, the scale matrix of the distribution.}
}
\value{
A numeric array of matrices, each of which is one simulation draw.
}
\description{
Simulate draws from the inverse Wishart distribution
}
\details{
Employs the Bartlett Decomposition (Smith & Hocking 1972).
Output exactly matches that of riwish from the MCMCpack package if the same
random seed is used.
}
