% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{solve_cubic}
\alias{solve_cubic}
\title{Find the roots of a * x^3 + b * x^2 + c * x + d = 0}
\usage{
solve_cubic(a, b, c, d)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector.}

\item{c}{Numeric vector.}

\item{d}{Numeric vector.}
}
\value{
Dataframe with four columns: \code{x1}, \code{x2}, \code{x3}, and
\code{three_real}. The first three contain the roots while the third
indicates whether all three roots are real, as described in the help file for
\code{solve_cubic_1}.
}
\description{
Find the roots of a * x^3 + b * x^2 + c * x + d = 0
}
\details{
This function is a wrapper to \code{solve_cubic_1} that allows the
coefficient on the cubic term to take on an arbitrary value.
}

