% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unusedFunctions.R
\name{solve_cubic_1}
\alias{solve_cubic_1}
\title{Find the roots of x^3 + a * x^2 + b * x + c = 0}
\usage{
solve_cubic_1(a, b, c)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector.}

\item{c}{Numeric vector.}
}
\value{
Dataframe with four columns: \code{x1}, \code{x2}, \code{x3},
and \code{three_real}. The column \code{x1} is real-valued since one root of
a cubic is necessarily real. The columns \code{x2} and \code{x3} are
complex-valued. Finally, \code{three_real} indicates whether the three roots
are all real-valued, i.e. if the imaginary parts of \code{x2} and \code{x3}
are both zero.
}
\description{
Find the roots of x^3 + a * x^2 + b * x + c = 0
}
\details{
\code{solve_cubic_1} is not intended to be called directly: it is
used by wrapper function \code{solve_cubic} which calculates the roots of an
arbitrary cubic equation, i.e. one with no restriction on the first
coefficient. Unlike \code{polyroot}, \code{solve_cubic}only accepts
real-valued coefficients, is fully vectorized, and uses an explicit rather
than iterative solution method.
}
