% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unusedFunctions.R
\name{solve_cubic_real}
\alias{solve_cubic_real}
\title{Find the real roots of a * x^3 + b * x^2 + c * x + d = 0}
\usage{
solve_cubic_real(a, b, c, d)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector.}

\item{c}{Numeric vector.}

\item{d}{Numeric vector.}
}
\value{
Dataframe containing the roots if real, NA otherwise.
}
\description{
Find the real roots of a * x^3 + b * x^2 + c * x + d = 0
}
