% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{solve_quadratic}
\alias{solve_quadratic}
\title{Find the roots of a * x^2 + b * x + c = 0}
\usage{
solve_quadratic(a, b, c)
}
\arguments{
\item{a}{Numeric vector}

\item{b}{Numeric vector}

\item{c}{Numeric vector}
}
\value{
Dataframe with columns x1, x2 and real. The first two columns contain
the complex-valued roots. The third column indicates if the roots are real.
}
\description{
Find the roots of a * x^2 + b * x + c = 0
}
\details{
This function computes an explicit solution unlike the base R
function \code{polyroot} which uses an iterative approach. Unlike
\code{polyroot}, \code{solve_quadratic} is vectorized in its arguments, but
only accepts real-valued coefficients as inputs.
}

