% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{solve_quadratic_real}
\alias{solve_quadratic_real}
\title{Find the real roots of a * x^2 + b * x + c = 0}
\usage{
solve_quadratic_real(a, b, c)
}
\arguments{
\item{a}{Numeric Vector.}

\item{b}{Numeric Vector.}

\item{c}{Numeric Vector.}
}
\value{
Dataframe with two columns, containing the roots if real and NA
otherwise.
}
\description{
Find the real roots of a * x^2 + b * x + c = 0
}
\details{
\code{solve_quadratic_real} simply calls \code{solve_quadratic} and
processes the results to remove any complex-valued roots. As such, it too is
vectorized in its arguments.
}

