% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{toList}
\alias{toList}
\title{Convert 3-d array to list of matrixes}
\usage{
toList(myArray)
}
\arguments{
\item{myArray}{A three-dimensional numeric array.}
}
\value{
A list of numeric matrices.
}
\description{
Convert 3-d array to list of matrixes
}
\examples{
M <- array(c(1, 1, 1, 1, 2, 2, 2, 2), c(2, 2, 2))
toList(M)
}

