% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{weber}
\alias{weber}
\title{Becker & Woessmann (2009) Dataset}
\format{\preformatted{'data.frame':	452 obs. of  44 variables:
 $ kreiskey1871       : int  1 2 3 4 5 6 7 8 9 10 ...
 $ county1871         : chr  "Memel" "Fischhausen" "Stadt K\\xf6nigsberg" "Landkreis K\\xf6nigsberg" ...
 $ rbkey              : int  1 1 1 1 1 1 1 1 1 1 ...
 $ lat_rad            : num  97.2 95.5 95.5 95.5 95.8 ...
 $ lon_rad            : num  36.9 34.9 35.8 35.8 36.8 ...
 $ kmwittenberg       : num  703 586 611 611 654 ...
 $ zupreussen         : int  1525 1525 1525 1525 1525 1525 1525 1525 1525 1525 ...
 $ hhsize             : num  4.69 5.26 4.65 5.37 4.98 ...
 $ gpop               : num  -0.798 -0.034 5.453 -1.17 0.406 ...
 $ f_prot             : num  92.9 98.3 92.7 99.4 99.1 ...
 $ f_jew              : num  2.012 0.342 3.422 0.149 0.478 ...
 $ f_rw               : num  68.6 80.4 90.1 75.3 68.2 ...
 $ f_miss             : num  2.49 1.74 0.31 2.57 1.27 ...
 $ f_young            : num  23.1 24.2 18.5 26 25.3 ...
 $ f_fem              : num  53.2 50.8 52.5 51.7 52.4 ...
 $ f_ortsgeb          : num  46 38.3 47.1 33.6 47.2 ...
 $ f_pruss            : num  98.6 99.9 99.2 99.9 100 ...
 $ f_blind            : num  0.1394 0.0807 0.1481 0.0809 0.1835 ...
 $ f_deaf             : num  0.145 0.138 0.143 0.162 0.26 ...
 $ f_dumb             : num  0.143 0.155 0.108 0.222 0.227 ...
 $ f_urban            : num  33.97 11.43 100 0 8.56 ...
 $ lnpop              : num  10.9 10.8 11.6 10.8 10.8 ...
 $ lnkmb              : num  6.42 6.22 6.27 6.27 6.35 ...
 $ poland             : num  0 0 0 0 0 0 0 0 0 0 ...
 $ latlon             : num  35.9 33.4 34.2 34.2 35.3 ...
 $ f_over3km          : num  15.2838 7.147 0.0398 3.04 4.2969 ...
 $ f_mine             : num  2.47 2.3302 0.0432 0.1011 0 ...
 $ inctaxpc           : num  197.7 134.4 NA 128.9 93.7 ...
 $ perc_secB          : num  18.8 17.3 30 14.2 12.4 ...
 $ perc_secC          : num  12.38 5.99 16.67 3.71 3.67 ...
 $ perc_secBnC        : num  31.2 23.3 46.7 17.9 16 ...
 $ lnyteacher         : num  678 662 713 667 662 ...
 $ rhs                : num  0 0 1 1 0 0 0 0 0 0 ...
 $ yteacher           : num  878 750 1247 786 752 ...
 $ pop                : num  55958 47103 112092 48218 50672 ...
 $ kmb                : num  616 501 527 527 570 ...
 $ uni1517            : int  0 0 0 0 0 0 0 0 0 0 ...
 $ reichsstadt        : int  0 0 0 0 0 0 0 0 0 0 ...
 $ hansestadt         : int  0 0 1 1 0 0 0 0 0 0 ...
 $ f_cath             : num  4.253 0.764 3.366 0.371 0.215 ...
 $ sh_al_in_tot       : num  0.497 0.5 1 0.451 0.449 ...
 $ ncloisters1517_pkm2: num  0 0.00188 0.14691 0 0.00094 ...
 $ school1517         : num  0 0 1 0 0 0 0 0 0 0 ...
 $ dnpop1500          : num  0 0 0.00441 0 0 ...
 - attr(*, "datalabel")= chr ""
 - attr(*, "time.stamp")= chr "24 May 2012 11:03"
 - attr(*, "formats")= chr  "\%8.0g" "\%33s" "\%8.0g" "\%9.0g" ...
 - attr(*, "types")= int  252 33 251 254 254 254 252 254 254 254 ...
 - attr(*, "val.labels")= chr  "" "" "" "" ...
 - attr(*, "var.labels")= chr  "kreiskey1871" "county name 1871" "districtkey" "Latitude in radius *100" ...
 - attr(*, "expansion.fields")=List of 4
  ..$ : chr  "kreiskey1871" "destring" "Characters removed were:"
  ..$ : chr  "rbkey" "destring" "Characters removed were:"
  ..$ : chr  "_dta" "_lang_c" "default"
  ..$ : chr  "_dta" "_lang_list" "default"
 - attr(*, "version")= int 12
}}
\source{
\url{https://www.cesifo-group.de/ifoHome/facts/iPEHD-Ifo-Prussian-Economic-History-Database/publications.html}
}
\usage{
weber
}
\description{
This is some data!
}
\references{
\url{https://www.cesifo-group.de/ifoHome/facts/iPEHD-Ifo-Prussian-Economic-History-Database.html}
\url{http://qje.oxfordjournals.org/content/124/2/531.short}
}
\keyword{datasets}

