% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afghan}
\alias{afghan}
\title{Burde and Linden (2013, AEJ Applied) Dataset}
\format{
A data frame with 687 rows and 17 variables:
\describe{
   \item{enrolled}{Indicator if child is enrolled in formal school. Outcome.}
   \item{testscore}{Normalized test score}
   \item{buildschool}{Indicator if village is treated. Instrument.}
   \item{headchild}{Indicator if child is child of head of household}
   \item{nhh}{Number of household members}
   \item{female}{Female indicator}
   \item{age}{Child's age}
   \item{yrsvill}{Time family has lived in village}
   \item{farsi}{Indicator for speaking Farsi}
   \item{tajik}{Indicator for speaking Tajik}
   \item{farmers}{Indicator for if head of household is a farmer}
   \item{land}{Number of jeribs of land owned}
   \item{agehead}{Head of household age}
   \item{educhead}{Years of education for head of household}
   \item{sheep}{Number of sheep and goats owned}
   \item{chagcharan}{Indicator if village is in Chagcharan district}
   \item{distschool}{Distance to nearest non-community based school}
}
}
\source{
Provided by author.
}
\usage{
afghan
}
\description{
Replicates IV using controls from Table 2
}
\references{
\url{http://www.jstor.org/stable/3083335}
}
\keyword{datasets}
