% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{collapse_3d_array}
\alias{collapse_3d_array}
\title{Collapse 3-d array to matrix}
\usage{
collapse_3d_array(myarray)
}
\arguments{
\item{myarray}{A three-dimensional array.}
}
\value{
Matrix with the 3rd dimension appended as rows to the matrix
}
\description{
Collapse 3-d array to matrix
}
