% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_bounds.R
\name{get_psi_upper}
\alias{get_psi_upper}
\title{Computes the upper bound of psi for binary data}
\usage{
get_psi_upper(s2_T, p, kappa)
}
\arguments{
\item{s2_T}{Vector of s2_T draws from observables}

\item{p}{Treatment probability from binary data}

\item{kappa}{Vector of kappa, NOTE: kappa_tilde in the paper}
}
\value{
Vector of upper bounds for psi
}
\description{
Computes the upper bound of psi for binary data
}
